+++ Modules/Intent.Modules.AspNetCore.MultiTenancy/Templates/Swashbuckle/TenantHeaderOperationFilter/TenantHeaderOperationFilterTemplateRegistration.cs
@@ -13,7 +13,7 @@ using Intent.Templates;
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.TemplateRegistration.Custom", Version = "1.0")]
 
-namespace Intent.Modules.AspNetCore.MultiTenancy.Templates.Swagger.TenantHeaderOperationFilter
+namespace Intent.Modules.AspNetCore.MultiTenancy.Templates.Swashbuckle.TenantHeaderOperationFilter
 {
     [IntentManaged(Mode.Merge, Body = Mode.Merge, Signature = Mode.Fully)]
     public class TenantHeaderOperationFilterTemplateRegistration : ITemplateRegistration