+++ Modules/Intent.Modules.AspNetCore.MultiTenancy/Templates/TemplateExtensions.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Intent.Modules.AspNetCore.MultiTenancy.Templates.MultiTenancyConfiguration;
 using Intent.Modules.AspNetCore.MultiTenancy.Templates.MultiTenantStoreDbContext;
-using Intent.Modules.AspNetCore.MultiTenancy.Templates.Swagger.TenantHeaderOperationFilter;
+using Intent.Modules.AspNetCore.MultiTenancy.Templates.Swashbuckle.TenantHeaderOperationFilter;
 using Intent.Modules.AspNetCore.MultiTenancy.Templates.TenantExtendedInfo;
 using Intent.Modules.Common.Templates;
 using Intent.RoslynWeaver.Attributes;